Bug 수정 (skinnedEnemyHPHandler, SeeThrough) HP(byte -> int) 건전지, 열쇠 동기화

+++ Project Marchen/Assets/Scripts/Projectiles/GrenadeHandler.cs
@@ -12,7 +12,7 @@ public class GrenadeHandler : NetworkBehaviour
 
     [Header("Grenade damage")]
     [SerializeField]
-    byte damageAmount;
+    int damageAmount;
 
     //Thrown by info
     PlayerRef thrownByPlayerRef;
